Move generate query string local
